DynamicBatch2D drawing API matches SpriteBatch
