+++ Source/Demos/Demo.Batching/Game1.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Text;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework.Input;
@@ -22,24 +23,14 @@ namespace Demo.Batching
         // ReSharper disable once NotAccessedField.Local
         private readonly GraphicsDeviceManager _graphicsDeviceManager;
 
-        private Batch2D _batch;
+        private DynamicBatch2D _batch;
+        private StringBuilder _stringBuilder = new StringBuilder();
         private SpriteBatch _spriteBatch;
         private BitmapFont _bitmapFont;
         private Texture2D _spriteTexture1;
         private Texture2D _spriteTexture2;
         private Vector2 _spriteOrigin;
         private DefaultEffect2D _effect;
-
-//        // the polygon
-//        private VertexPositionColor[] _polygonVertices;
-//        private short[] _polygonIndices;
-//        // the curve (continous line segements)
-//        private VertexPositionColor[] _curveVertices;
-//
-//        // primitives matrix transformation chain
-//        private Matrix _primitivesModelToWorld;
-//        private Matrix _primitivesWorldToView;
-//        private Matrix _primitivesViewToProjetion;
 
         private readonly Random _random = new Random();
         private readonly FramesPerSecondCounter _fpsCounter = new FramesPerSecondCounter();
@@ -66,65 +57,15 @@ namespace Demo.Batching
         {
             var graphicsDevice = GraphicsDevice;
 
-            _batch = new Batch2D(graphicsDevice);
+            _batch = new DynamicBatch2D(graphicsDevice);
             _spriteBatch = new SpriteBatch(graphicsDevice);
             _bitmapFont = Content.Load<BitmapFont>("montserrat-32");
             _effect = new DefaultEffect2D(graphicsDevice);
-
-//            var viewport = graphicsDevice.Viewport;
-
-//            // the transformation used to transform primitives from their model space to the world space
-//            // here we scale the x, y and z axes by 100 units
-//            _primitivesModelToWorld = Matrix.CreateScale(scales: new Vector3(x: 100, y: 100, z: 100));
-//
-//            // the camera transformation used to transform primitives from world space to a view space
-//            // here we don't do anything by using the identity matrix
-//            _primitivesWorldToView = Matrix.Identity;
-//
-//            // the transformation used to transform primitives from view space to projection space
-//            // here we create an orthographic projection; a 3D box where any primitives outside this box are not rendered
-//            // here the box is created off an origin point (0,0,0) which is the centre of the screen's surface
-//            _primitivesViewToProjetion = Matrix.CreateOrthographicOffCenter(left: viewport.Width * -0.5f, right: viewport.Width * 0.5f, bottom: viewport.Height * -0.5f, top: viewport.Height * 0.5f, zNearPlane: 0, zFarPlane: 1);
 
             // load the texture for the sprites
             _spriteTexture1 = Content.Load<Texture2D>("logo-square-128");
             _spriteTexture2 = Content.Load<Texture2D>("logo-square-128-copy");
             _spriteOrigin = new Vector2(_spriteTexture1.Width * 0.5f, _spriteTexture1.Height * 0.5f);
-
-
-            //            // create our polygon mesh; vertices are in Local space; indices are index references to the vertices to draw 
-            //            // indices have to multiple of 3 for PrimitiveType.TriangleList which says to draw a collection of triangles each with 3 vertices (different triangles can share vertices) 
-            //            // here we have 2 triangles in the list to form a quad or rectangle: http://wiki.lwjgl.org/images/a/a8/QuadVertices.png
-            //            // TriangleList is the most common scenario to have polygon vertices layed out in memory for uploading to the GPU
-            //            _polygonVertices = new[]
-            //            {
-            //                new VertexPositionColor(position: new Vector3(x: 0, y: 0, z: 0), color: Color.Red),
-            //                new VertexPositionColor(position: new Vector3(x: 2, y: 0, z: 0), color: Color.Blue),
-            //                new VertexPositionColor(position: new Vector3(x: 1, y: 2, z: 0), color: Color.Green),
-            //                new VertexPositionColor(position: new Vector3(x: 3, y: 2, z: 0), color: Color.White)
-            //            };
-            //            _polygonIndices = new short[]
-            //            {
-            //                1,
-            //                0,
-            //                2,
-            //                1,
-            //                2,
-            //                3
-            //            };
-            //
-            //            // create our curve as an approximation by a series of line segments; vertices are in Local space; no indices
-            //            // LineStrip joins the vertices given in order into a continuous series of line segments
-            //            var curveVertices = new List<VertexPositionColor>();
-            //            var angleStep = MathHelper.ToRadians(degrees: 1);
-            //            const int circlesCount = 3;
-            //            for (var angle = 0.0f; angle <= MathHelper.TwoPi * circlesCount; angle += angleStep)
-            //            {
-            //                var vertexPosition = new Vector3(x: (float)Math.Sin(angle) - angle / 10, y: (float)Math.Cos(angle) - angle / 15, z: 0);
-            //                var vertex = new VertexPositionColor(vertexPosition, Color.White);
-            //                curveVertices.Add(vertex);
-            //            }
-            //            _curveVertices = curveVertices.ToArray();
 
             var viewport = GraphicsDevice.Viewport;
 
@@ -178,8 +119,7 @@ namespace Demo.Batching
         {
             var graphicsDevice = GraphicsDevice;
 
-            // clear the (pixel) buffer to a specific color
-            graphicsDevice.Clear(Color.CornflowerBlue);
+            graphicsDevice.Clear(Color.Black);
 
             // set the states for rendering
             // this could be moved outside the render loop if it doesn't change frame per frame 
@@ -190,12 +130,26 @@ namespace Demo.Batching
             graphicsDevice.DepthStencilState = DepthStencilState.None;
             graphicsDevice.RasterizerState = RasterizerState.CullNone;
 
-            // draw the polygon and curve in the cartesian coordinate system using the VertexPositionColor PrimitiveBatch
-            //            _primitiveBatchPositionColor.Begin();
-            //            _primitiveBatchPositionColor.Draw(_primitiveMaterial, PrimitiveType.TriangleList, _polygonVertices, _polygonIndices);
-            //            _primitiveBatchPositionColor.Draw(_primitiveMaterial, PrimitiveType.LineStrip, _curveVertices);
-            //            _primitiveBatchPositionColor.End();
+            // comment and uncomment either of the two below lines to compare
+            DrawSpritesWithBatch2D();
+            //DrawSpritesWithSpriteBatch();
 
+            _batch.Begin(sortMode: Batch2DSortMode.Texture, effect: _effect);
+
+            _stringBuilder.Clear();
+            _stringBuilder.Append("FPS: ");
+            _stringBuilder.Append(_fpsCounter.FramesPerSecond);
+            _batch.DrawString(_bitmapFont, _stringBuilder, Vector2.Zero);
+
+            _batch.End();
+
+            base.Draw(gameTime);
+
+            _fpsCounter.Draw(gameTime);
+        }
+
+        private void DrawSpritesWithBatch2D()
+        {
             _batch.Begin(sortMode: Batch2DSortMode.Texture, effect: _effect);
 
             // ReSharper disable once ForCanBeConvertedToForeach
@@ -206,22 +160,20 @@ namespace Demo.Batching
             }
 
             _batch.End();
+        }
 
-            //_spriteBatch.Begin(sortMode: SpriteSortMode.Texture, effect: _effect);
+        private void DrawSpritesWithSpriteBatch()
+        {
+            _spriteBatch.Begin(sortMode: SpriteSortMode.Texture, blendState: BlendState.AlphaBlend, samplerState: SamplerState.PointClamp, effect: _effect);
 
-            //// ReSharper disable once ForCanBeConvertedToForeach
-            //for (var index = 0; index < _sprites.Length; index++)
-            //{
-            //    var sprite = _sprites[index];
-            //    _spriteBatch.Draw(sprite.Texture, sprite.Position, rotation: sprite.Rotation, origin: _spriteOrigin, color: sprite.Color);
-            //}
+            // ReSharper disable once ForCanBeConvertedToForeach
+            for (var index = 0; index < _sprites.Length; index++)
+            {
+                var sprite = _sprites[index];
+                _spriteBatch.Draw(sprite.Texture, sprite.Position, rotation: sprite.Rotation, origin: _spriteOrigin, color: sprite.Color);
+            }
 
-            //_spriteBatch.End();
-
-            base.Draw(gameTime);
-
-            _fpsCounter.Draw(gameTime);
-            Window.Title = $"Demo.Batching - FPS: {_fpsCounter.FramesPerSecond}";
+            _spriteBatch.End();
         }
     }
 }
