+++ Source/MonoGame.Extended/Graphics/Batch2DSortMode.cs
@@ -16,7 +16,7 @@ namespace MonoGame.Extended.Graphics
 
         /// <summary>
         ///     Each draw call is added to the end of an array of draw commands. Draw commands will be merged, if possible, for
-        ///     batching. When <see cref="Batch2D.End" /> is called, the merged draw commands are processed on a first come, first
+        ///     batching. When <see cref="DynamicBatch2D.End" /> is called, the merged draw commands are processed on a first come, first
         ///     serve, basis where each command invokes
         ///     <see cref="GraphicsDevice.DrawIndexedPrimitives(PrimitiveType, int, int, int)" /> for each
         ///     <see cref="EffectPass" />.
@@ -26,7 +26,7 @@ namespace MonoGame.Extended.Graphics
         /// <summary>
         ///     Each draw call is added to the end of an array of draw commands. Draw commands will be merged, if possible, for
         ///     batching as they are enqueued and after they are sorted by their <see cref="Texture2D" />s. When
-        ///     <see cref="Batch2D.End" /> is called, the sorting takes place and then the merged draw commands are processed where
+        ///     <see cref="DynamicBatch2D.End" /> is called, the sorting takes place and then the merged draw commands are processed where
         ///     each command invokes <see cref="GraphicsDevice.DrawIndexedPrimitives(PrimitiveType, int, int, int)" /> for each
         ///     <see cref="EffectPass" />.
         /// </summary>
@@ -35,7 +35,7 @@ namespace MonoGame.Extended.Graphics
         /// <summary>
         ///     Each draw call is added to the end of an array of draw commands. Draw commands will be merged, if possible, for
         ///     batching as they are enqueued and after they are sorted by their depth values in ascending order. When
-        ///     <see cref="Batch2D.End" /> is called, the sorting takes place and then the merged draw commands are processed where
+        ///     <see cref="DynamicBatch2D.End" /> is called, the sorting takes place and then the merged draw commands are processed where
         ///     each command invokes <see cref="GraphicsDevice.DrawIndexedPrimitives(PrimitiveType, int, int, int)" /> for each
         ///     <see cref="EffectPass" />.
         /// </summary>
@@ -50,7 +50,7 @@ namespace MonoGame.Extended.Graphics
         /// <summary>
         ///     Each draw call is added to the end of an array of draw commands. Draw commands will be merged, if possible, for
         ///     batching as they are enqueued and after they are sorted by their depth values in descending order. When
-        ///     <see cref="Batch2D.End" /> is called, the sorting takes place and then the merged draw commands are processed where
+        ///     <see cref="DynamicBatch2D.End" /> is called, the sorting takes place and then the merged draw commands are processed where
         ///     each command invokes <see cref="GraphicsDevice.DrawIndexedPrimitives(PrimitiveType, int, int, int)" /> for each
         ///     <see cref="EffectPass" />.
         /// </summary>
