+++ Source/MonoGame.Extended/MonoGame.Extended.csproj
@@ -90,7 +90,7 @@
     <Compile Include="Graphics\GeometryBufferExtensions.VertexPositionColorTexture.cs" />
     <Compile Include="Graphics\GeometryBufferOverflowException.cs" />
     <Compile Include="Graphics\PrimitiveTypeExtensions.cs" />
-    <Compile Include="Graphics\Batch2D.cs" />
+    <Compile Include="Graphics\DynamicBatch2D.cs" />
     <Compile Include="Graphics\GeometryBuffer.Static.cs" />
     <Compile Include="Gui\Controls\GuiLabel.cs" />
     <Compile Include="Gui\Controls\GuiToggleButton.cs" />
